refactor: update type exports and barrel files for feature structure
