+++ src/features/index.ts
@@ -5,3 +5,4 @@ export * from "./providers";
 export * from "./dashboard";
 export * from "./auth";
 export * from "./email";
+export * from "./shared";
