+++ src/features/providers/types/index.ts
@@ -1,2 +1,3 @@
 // providers types exports
 export * from "./provider";
+export * from "./provider-form-data";
