+++ src/features/tenants/types/index.ts
@@ -1,2 +1,3 @@
 // tenants types exports
 export * from "./tenant";
+export * from "./tenant-form-data";
