+++ src/types/index.ts
@@ -1,6 +1,10 @@
-export * from "@/types/bill";
-export * from "@/types/email-content";
-export * from "@/types/form-data";
-export * from "@/types/tenant";
-export * from "@/types/user";
-export * from "@/types/utility-provider";
+// re-export from feature-based structure
+export * from "@/features/bills/types";
+export * from "@/features/email/types";
+export * from "@/features/tenants/types";
+export * from "@/features/auth/types";
+export * from "@/features/providers/types";
+export * from "@/features/shared";
+
+// legacy exports for backward compatibility
+// Note: form-data types have been moved to their respective features
